Add files via upload
This is a fun mini project that I built in Python — “Guess the Number Game.”
In this game, the computer randomly selects a number between 1 to 10, and the player has 3 attempts to guess the correct number. If the correct number isn’t guessed within those attempts, the player loses. Simple, exciting, and full of logic! 

This exercise uses some important core concepts:
🔹 random module
🔹 Conditional statements (if-elif-else)
🔹 Looping
🔹 User input handling
🔹 break statement for loop control